+++ ios/Classes/FlutterSystemProxyPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FlutterSystemProxyPlugin : NSObject<FlutterPlugin>
-@end