+++ ios/flutter_system_proxy/Package.swift
@@ -1,0 +1,24 @@
+// swift-tools-version: 5.9
+import PackageDescription
+
+let package = Package(
+    name: "flutter_system_proxy",
+    platforms: [.iOS("12.0")],
+    products: [
+        .library(name: "flutter-system-proxy", targets: ["flutter_system_proxy"])
+    ],
+    dependencies: [
+        .package(name: "FlutterFramework", path: "../FlutterFramework")
+    ],
+    targets: [
+        .target(
+            name: "flutter_system_proxy",
+            dependencies: [
+                .product(name: "FlutterFramework", package: "FlutterFramework")
+            ],
+            linkerSettings: [
+                .linkedFramework("CFNetwork")
+            ]
+        )
+    ]
+)